+++ crates/data/src/engine/mod.rs
@@ -56,7 +56,7 @@ use nautilus_common::{
     clock::Clock,
     logging::{RECV, RES},
     messages::data::{
-        DataCommand, DataRequest, DataResponse, SubscribeBars, SubscribeBookDeltas,
+        DataCommand, DataResponse, RequestCommand, SubscribeBars, SubscribeBookDeltas,
         SubscribeBookDepth10, SubscribeBookSnapshots, SubscribeCommand, UnsubscribeBars,
         UnsubscribeBookDeltas, UnsubscribeBookDepth10, UnsubscribeBookSnapshots,
         UnsubscribeCommand,
@@ -216,17 +216,7 @@ impl DataEngine {
         subs
     }
 
-    fn get_client(&self, client_id: &ClientId, venue: &Venue) -> Option<&DataClientAdapter> {
-        match self.clients.get(client_id) {
-            Some(client) => Some(client),
-            None => self
-                .routing_map
-                .get(venue)
-                .and_then(|client_id: &ClientId| self.clients.get(client_id)),
-        }
-    }
-
-    fn get_client_mut(
+    fn get_client(
         &mut self,
         client_id: Option<&ClientId>,
         venue: Option<&Venue>,
@@ -362,10 +352,7 @@ impl DataEngine {
         let result = match cmd {
             DataCommand::Subscribe(cmd) => self.execute_subscribe(cmd),
             DataCommand::Unsubscribe(cmd) => self.execute_unsubscribe(cmd),
-            DataCommand::Request(cmd) => {
-                self.execute_request(cmd);
-                Ok(())
-            }
+            DataCommand::Request(cmd) => self.execute_request(cmd),
         };
 
         if let Err(e) = result {
@@ -382,7 +369,7 @@ impl DataEngine {
             _ => {} // Do nothing else
         }
 
-        if let Some(client) = self.get_client_mut(cmd.client_id(), cmd.venue()) {
+        if let Some(client) = self.get_client(cmd.client_id(), cmd.venue()) {
             client.execute_subscribe_command(cmd.clone());
         } else {
             log::error!(
@@ -404,7 +391,7 @@ impl DataEngine {
             _ => {} // Do nothing else
         }
 
-        if let Some(client) = self.get_client_mut(cmd.client_id(), cmd.venue()) {
+        if let Some(client) = self.get_client(cmd.client_id(), cmd.venue()) {
             client.execute_unsubscribe_command(cmd.clone());
         } else {
             log::error!(
@@ -417,14 +404,23 @@ impl DataEngine {
         Ok(())
     }
 
-    /// Sends a [`DataRequest`] to an endpoint that must be a data client implementation.
-    pub fn execute_request(&self, req: DataRequest) {
-        if let Some(client) = self.get_client(&req.client_id, &req.venue) {
-            client.through_request(req);
+    /// Sends a [`RequestCommand`] to an endpoint that must be a data client implementation.
+    pub fn execute_request(&mut self, req: RequestCommand) -> anyhow::Result<()> {
+        if let Some(client) = self.get_client(req.client_id(), req.venue()) {
+            match req {
+                RequestCommand::Data(req) => client.request_data(req),
+                RequestCommand::Instrument(req) => client.request_instrument(req),
+                RequestCommand::Instruments(req) => client.request_instruments(req),
+                RequestCommand::BookSnapshot(req) => client.request_book_snapshot(req),
+                RequestCommand::Quotes(req) => client.request_quotes(req),
+                RequestCommand::Trades(req) => client.request_trades(req),
+                RequestCommand::Bars(req) => client.request_bars(req),
+            }
         } else {
-            log::error!(
-                "Cannot handle request: no client found for {}",
-                req.client_id
+            anyhow::bail!(
+                "Cannot handle request: no client found for {:?} {:?}",
+                req.client_id(),
+                req.venue()
             );
         }
     }