+++ crates/pyo3/src/lib.rs
@@ -111,8 +111,8 @@ fn nautilus_pyo3(py: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     sys_modules.set_item(format!("{module_name}.{n}"), m.getattr(n)?)?;
     re_export_module_attributes(m, n)?;
 
-    let n = "test_kit";
-    let submodule = pyo3::wrap_pymodule!(nautilus_test_kit::python::test_kit);
+    let n = "testkit";
+    let submodule = pyo3::wrap_pymodule!(nautilus_testkit::python::testkit);
     m.add_wrapped(submodule)?;
     sys_modules.set_item(format!("{module_name}.{n}"), m.getattr(n)?)?;
     re_export_module_attributes(m, n)?;